+++ turing-app/src/test/java/com/viglet/turing/api/sn/graphql/GraphQLConfigTest.java
@@ -1,8 +1,13 @@
 package com.viglet.turing.api.sn.graphql;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.mockito.Mockito.mock;
 
 import org.junit.jupiter.api.Test;
+
+import com.viglet.turing.persistence.repository.sn.TurSNSiteRepository;
+import com.viglet.turing.persistence.repository.sn.field.TurSNSiteFieldExtRepository;
 
 import graphql.schema.idl.RuntimeWiring;
 
@@ -10,9 +15,12 @@ class GraphQLConfigTest {
 
     @Test
     void testConfigure() {
-        GraphQLConfig config = new GraphQLConfig();
+        TurSNSiteRepository siteRepository = mock(TurSNSiteRepository.class);
+        TurSNSiteFieldExtRepository fieldExtRepository = mock(TurSNSiteFieldExtRepository.class);
+        GraphQLConfig config = new GraphQLConfig(siteRepository, fieldExtRepository);
         RuntimeWiring.Builder builder = RuntimeWiring.newRuntimeWiring();
 
         assertDoesNotThrow(() -> config.configure(builder));
+        assertNotNull(config.dynamicSearchDocumentFieldsCustomizer());
     }
 }